requirements.txt
# ============================================
# Crypto Screener - Python Dependencies
# ============================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# Exchange Integration
ccxt==4.2.25

# Telegram
python-telegram-bot==20.7

# Database
aiosqlite==0.19.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
aiofiles==23.2.1

# For production (optional)
# gunicorn==21.2.0
